requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1733332318622/work
argcomplete==3.5.3
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attr==0.3.1
attrs==24.3.0
azure-core==1.32.0
azure-storage-blob==12.24.0
black==24.10.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.35.92
botocore==1.35.92
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.40.0
datamodel-code-generator==0.26.1
debugpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_563_nwtkoc/croot/debugpy_1690905063850/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
Django==4.2.17
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
docker==7.1.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
expiringdict==1.2.2
Faker==33.3.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
idna==3.7
ijson==3.3.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1733223117029/work
inflect==5.6.2
inflection==0.5.1
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1734788142186/work
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.8
label-studio==1.15.0
label-studio-sdk==1.0.8
launchdarkly-server-sdk==8.2.1
lightning==2.5.0.post0
lightning-utilities==0.11.9
lockfile==0.12.2
lxml==5.3.0
lxml_html_clean==0.4.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mlflow==2.19.0
mlflow-skinny==2.19.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.59.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
ordered-set==4.0.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pathspec==0.12.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1733302527033/work
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil @ file:///Users/cbousseau/work/recipes/ci_py311_2/psutil_1678995687212/work
psycopg2-binary==2.9.10
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1733221634316/work
pyparsing==3.2.1
pyRFC3339==2.0.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
python-json-logger==2.0.4
pytorch-lightning==2.5.0.post0
pytz==2022.7.1
PyYAML==6.0.2
pyzmq @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_95lsut8ymz/croot/pyzmq_1734709560733/work
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.50
rpds-py==0.22.3
rq==1.10.1
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
semver==3.0.2
sentry-sdk==2.19.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.1
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tornado @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0axef5a0m0/croot/tornado_1733960501260/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2024.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.57
ultralytics-thop==2.0.13
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.0
xmljson==0.2.1
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
